Add basic definition for MaterialImage struct
MaterialImage has a vector for the local states for all cells, a size
tuple for measurements, and an integer to take note of the required
number of discrete states to study

Add constructors and overload Base functions

Add two constructors for MaterialImage from two arrays. First, from a vector given the number of discs and
2d shape. Second, from a 2d array and the number of d
iscs.

Overload size to show the dimensions of the type. Ove
rload length to show the length of the vector in the
type. Overload unique to return the number of expecte
d discs. Overload eltype to show the type of floats s
tored.

Create function segmented to return whether the image
 has been segmented or not.